made profile view show field data
